fix footer and navbar small screen display:
