+++ src/app/page.jsx
@@ -1,11 +1,9 @@
-import NavBar from "./ui/Navbar";
 import Footer from "./ui/Footer";
 import Main from "./ui/Main";
 
 export default function Home() {
   return (
     <>
-      <NavBar />
       <Main />
       <Footer />
     </>
