+++ src/app/ui/BestSellerCarousel.jsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import Link from "next/link";
 import Dropdown from 'react-bootstrap/Dropdown';
-import data from "./data";
+import data from "@/utils/data";
 import useScreenSize from "@/components/screenSize";
 import { FaRegSmileBeam } from "react-icons/fa";
