+++ src/app/ui/FavoriteCarrousel.jsx
@@ -1,6 +1,6 @@
 'use client'
 
-import data from "./data";
+import data from "@/utils/data";
 import Image from "next/image";
 
 export default function FavoriteCarrousel() {
