+++ src/app/ui/Main.jsx
@@ -6,6 +6,7 @@ export default function Main() {
     <div className="container-fluid p-0">
       <FavoriteCarrousel />
       <BestSellerCarousel />
+      
     </div>
   );
 }
