+++ src/components/provider.jsx
@@ -1,0 +1,14 @@
+'use client';
+
+import SessionProvider from 'next-auth/react';
+
+const Provider = ({children, session}) => {
+  return (
+    <SessionProvider session={session}>
+      {children}
+    </ SessionProvider>
+  )
+
+}
+
+export default Provider;
